Move clientSession.
